CS-10585: Consolidate software-factory module layout (src/ vs scripts/lib/)
Move all 20 library modules from scripts/lib/ into src/ so there is a
single canonical location for importable code. Remove the obsolete
smoke-realm CLI entrypoint and relocate smoke-test-realm into
scripts/smoke-tests/ where it belongs. Update every import path and
__dirname-relative resolution across src/, scripts/, and tests/.

All 386 node tests pass.
