+++ packages/software-factory/package.json
@@ -5,7 +5,6 @@
   "license": "MIT",
   "description": "Software Factory workspace package",
   "scripts": {
-    "boxel:pick-ticket": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/pick-ticket.ts",
     "boxel:search": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/boxel-search.ts",
     "boxel:session": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/boxel-session.ts",
     "cache:prepare": "NODE_NO_WARNINGS=1 ts-node --transpileOnly src/cli/cache-realm.ts",
@@ -25,8 +24,7 @@
     "lint:types": "ember-tsc --noEmit",
     "serve:realm": "NODE_NO_WARNINGS=1 ts-node --transpileOnly src/cli/serve-realm.ts",
     "serve:support": "NODE_NO_WARNINGS=1 ts-node --transpileOnly src/cli/serve-support.ts",
-    "smoke:realm": "NODE_NO_WARNINGS=1 ts-node --transpileOnly src/cli/smoke-realm.ts",
-    "smoke:test-realm": "NODE_NO_WARNINGS=1 ts-node --transpileOnly src/cli/smoke-test-realm.ts",
+    "smoke:test-realm": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/smoke-tests/smoke-test-realm.ts",
     "test": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/test.ts",
     "test:all": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/test.ts",
     "test:node": "NODE_NO_WARNINGS=1 ts-node --transpileOnly scripts/test.ts --node-only",
