+++ packages/software-factory/scripts/smoke-tests/factory-loop-smoke.ts
@@ -21,9 +21,9 @@ import type {
   ProjectCard,
   TestResult,
   TicketCard,
-} from '../lib/factory-agent';
+} from '../../src/factory-agent';
 
-import type { FactoryTool, ToolCallEntry } from '../lib/factory-tool-builder';
+import type { FactoryTool, ToolCallEntry } from '../../src/factory-tool-builder';
 
 import {
   runFactoryLoop,
@@ -34,7 +34,7 @@ import {
   type FactoryLoopResult,
   type LoopAgent,
   type TestRunner,
-} from '../lib/factory-loop';
+} from '../../src/factory-loop';
 
 // ---------------------------------------------------------------------------
 // Helpers
