+++ packages/software-factory/scripts/smoke-tests/factory-skill-smoke.ts
@@ -20,8 +20,8 @@ import {
   SkillLoader,
   enforceSkillBudget,
   estimateTokens,
-} from '../lib/factory-skill-loader';
-import type { ProjectCard, TicketCard } from '../lib/factory-agent';
+} from '../../src/factory-skill-loader';
+import type { ProjectCard, TicketCard } from '../../src/factory-agent';
 
 const SAMPLE_TICKETS: { label: string; ticket: TicketCard }[] = [
   {
