+++ packages/software-factory/src/factory-agent-mocks.ts
