+++ packages/software-factory/src/factory-agent-tool-use.ts
@@ -12,7 +12,7 @@ import { SupportedMimeType } from '@cardstack/runtime-common/supported-mime-type
 
 const MAX_TOOL_USE_TURNS = 50;
 
-import { createBoxelRealmFetch } from '../../src/realm-auth';
+import { createBoxelRealmFetch } from './realm-auth';
 
 import type {
   AgentContext,
