+++ packages/software-factory/src/factory-agent-types.ts
