+++ packages/software-factory/src/factory-context-builder.ts
