+++ packages/software-factory/src/factory-tool-registry.ts
