+++ packages/software-factory/tests/factory-tool-executor.spec.ts
@@ -16,10 +16,10 @@ import { expect } from '@playwright/test';
 import {
   ToolExecutor,
   ToolNotFoundError,
-} from '../scripts/lib/factory-tool-executor';
-import { ToolRegistry } from '../scripts/lib/factory-tool-registry';
-import { buildFactoryTools } from '../scripts/lib/factory-tool-builder';
-import { fetchCardTypeSchema } from '../scripts/lib/darkfactory-schemas';
+} from '../src/factory-tool-executor';
+import { ToolRegistry } from '../src/factory-tool-registry';
+import { buildFactoryTools } from '../src/factory-tool-builder';
+import { fetchCardTypeSchema } from '../src/darkfactory-schemas';
 import {
   baseRealmURLFor,
   DEFAULT_REALM_OWNER,
@@ -197,7 +197,7 @@ test('unregistered tool is rejected without reaching the server', async ({
 // Factory tool (card write) tests against live realm
 // ---------------------------------------------------------------------------
 
-import type { FactoryTool } from '../scripts/lib/factory-tool-builder';
+import type { FactoryTool } from '../src/factory-tool-builder';
 
 type CardWriteResult = { ok: boolean; error?: string };
 type CardReadResult = { ok: boolean; document?: LooseSingleCardDocument };
