+++ packages/software-factory/tests/runtime-schema.spec.ts
@@ -16,8 +16,8 @@ import { expect } from '@playwright/test';
 import {
   runRealmCommand,
   ensureTrailingSlash,
-} from '../scripts/lib/realm-operations';
-import { fetchCardTypeSchema } from '../scripts/lib/darkfactory-schemas';
+} from '../src/realm-operations';
+import { fetchCardTypeSchema } from '../src/darkfactory-schemas';
 import { sourceRealmURLFor } from '../src/harness/shared';
 
 const GET_CARD_TYPE_SCHEMA_COMMAND =
